requirements.txt
click
Flask~=2.3.2
itsdangerous
Jinja2
MarkupSafe
mysql-connector-python~=8.0.33
protobuf
python-slugify~=8.0.1
six
text-unidecode
timeago~=1.0.16
Werkzeug
flask_sitemapper
cryptography~=41.0.1
geoip2
requests~=2.31.0
selenium~=4.10.0
urllib3~=2.0.3
bs4~=0.0.1
beautifulsoup4~=4.12.2